move commands directory under controllers
I felt commands didn't make sense as a top-level directory. Since all of
the commands would be used in the controllers, I moved it as a
sub-directory under controllers/
